refactor. util 내 date.js -> util.js 통합
